ensure styles consistency, leverage mui props, organize components
