+++ jsconfig.json
@@ -7,8 +7,7 @@
       "@/hooks/*": ["hooks/*"],
       "@/icons/*": ["icons/*"],
       "@/lib/*": ["lib/*"],
-      "@/services/*": ["services/*"],
-      "@/styles/*": ["archive/styles/*"]
+      "@/services/*": ["services/*"]
     }
   }
 }
