+++ package.json
@@ -15,13 +15,13 @@
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
-    "@prisma/cli": "2.10.1",
-    "@prisma/client": "2.10.1",
+    "@prisma/cli": "2.11.0",
+    "@prisma/client": "2.11.0",
     "@welldone-software/why-did-you-render": "^5.0.0",
     "clsx": "^1.1.1",
     "dotenv": "^8.2.0",
-    "eslint": "7.12.1",
-    "eslint-config-airbnb": "^18.2.0",
+    "eslint": "7.13.0",
+    "eslint-config-airbnb": "^18.2.1",
     "eslint-import-resolver-alias": "^1.1.2",
     "eslint-plugin-cypress": "^2.11.2",
     "eslint-plugin-import": "^2.22.1",
@@ -34,7 +34,7 @@
     "isomorphic-unfetch": "^3.1.0",
     "moment": "^2.29.1",
     "next": "latest",
-    "next-compose-plugins": "^2.2.0",
+    "next-compose-plugins": "^2.2.1",
     "next-svgr": "0.0.2",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
