+++ src/components/dropdown.js
@@ -3,12 +3,10 @@ import { makeStyles } from '@material-ui/core/styles';
 import Menu from '@material-ui/core/Menu';
 import theme from '@/lib/theme';
 
-const useStyles = makeStyles({
+const menuStyles = makeStyles({
   paper: {
     border: `1px solid ${theme.palette.grays.default}`,
-    borderRadius: 0,
-    boxShadow: 'none',
-    marginTop: theme.spacing(1),
+    borderRadius: 0
   },
   list: {
     padding: `${theme.spacing(1)}px ${theme.spacing(2)}px`,
@@ -19,14 +17,15 @@ const CustomMenu = ({
   children,
   ...rest
 }) => {
-  const classes = useStyles();
+  const menuClasses = menuStyles();
 
   return (
     <Menu
       classes={{
-        paper: classes.paper,
-        list: classes.list,
+        paper: menuClasses.paper,
+        list: menuClasses.list,
       }}
+      elevation={0}
       getContentAnchorEl={null}
       keepMounted
       {...rest}
