--- src/hooks/user.js
@@ -1,5 +1,0 @@
-export const UserName = (user) => user && `${user.nameFirst} ${user.nameLast}`;
-
-export const UserNameId = (user) => (
-  user && `${user.nameFirst.toLowerCase()}-${user.nameLast.toLowerCase()}`
-);
