+++ src/pages/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Layout from '@/components/layout';
-import withAuth from '@/components/auth/with-auth';
+import withAuth from 'auth/with-auth';
 
 const Home = () => (
   <Layout title="Home">
